+++ src/index.ts
@@ -1,6 +1,5 @@
 export { ApiError } from "./utils/ApiError.js";
 export { isEqual } from "./utils/isEqual.js";
-export { makeModel } from "./schema/model";
 export { Schema } from "./schema";
 export * as validate from "./validate";
 export * from "./utils/interfaces";